update department operators

+++ server/src/resolvers/departmentUser.js
@@ -1,6 +1,6 @@
 import { combineResolvers } from 'graphql-resolvers'
-import { UserInputError } from 'apollo-server'
-import { isDepartmentMember } from './authorization'
+import { UserInputError, ApolloError } from 'apollo-server'
+import { isDepartmentMember, isDepartmentOperator } from './authorization'
 
 const toCursorHash = (string) => Buffer.from(string).toString('base64')
 
@@ -8,17 +8,6 @@ const fromCursorHash = (string) => Buffer.from(string, 'base64').toString('ascii
 
 export default {
   Query: {
-    // TODO: authentication
-    usersOfDepartment: combineResolvers(
-      isDepartmentMember,
-      async (root, { departmentId }, { models }) => {
-        const dpmUsers = await models.DepartmentUser.find({ departmentId }).populate('userId')
-        const users = dpmUsers.map((item) => item.userId)
-        // console.log('users: ', users)
-
-        return users
-      }
-    ),
     departmentUsers: combineResolvers(
       async (root, { cursor, limit = 50, userId, departmentId, role }, { models }) => {
         if (role) {
@@ -32,36 +21,42 @@ export default {
   },
 
   Mutation: {
-    // TODO: only allow head of department to invite user
-    inviteMember: async (root, { departmentId, email, role }, { models, newErr }) => {
-      const userFound = await models.User.findOne({ email })
-      if (!userFound) {
-        throw new UserInputError('Email không hợp lệ')
-      }
-      const validRoles = ['reviewer', 'member']
+    inviteMember: combineResolvers(
+      isDepartmentOperator,
+      async (root, { departmentId, email, role }, { models, newErr }) => {
+        const userFound = await models.User.findOne({ email })
+        if (!userFound) {
+          throw new UserInputError('Email is not valid')
+        }
+        const validRoles = ['chief', 'reviewer', 'member']
 
-      if (!validRoles.includes(role)) {
-        throw new UserInputError('Vai trò thành viên không hợp lệ')
-      }
+        if (!validRoles.includes(role)) {
+          throw new UserInputError('Role specification is not valid')
+        }
 
-      const isAlreadyMember = await models.DepartmentUser.findOne({
-        userId: userFound.id,
-        departmentId
-      })
+        const isAlreadyMember = await models.DepartmentUser.findOne({
+          userId: userFound.id,
+          departmentId
+        })
 
-      if (isAlreadyMember) {
-        throw new UserInputError('Người dùng này đã là thành viên của khoa')
+        if (isAlreadyMember) {
+          throw new UserInputError('This user is already a member')
+        }
+        const document = await models.DepartmentUser.create({
+          userId: userFound.id,
+          departmentId,
+          departmentRole: role
+        })
+        return document
       }
-      const document = await models.DepartmentUser.create({
-        userId: userFound.id,
-        departmentId,
-        departmentRole: role
-      })
-      return document
-    },
+    ),
     removeMember: combineResolvers(
       // TODO: handle only head of department is allowed
-      async (root, { departmentId, userId }, { models, newErr }) => {
+      isDepartmentOperator,
+      async (root, { departmentId, userId }, { models, me, newErr }) => {
+        if (userId === me.id) {
+          throw new ApolloError('Cannot remove self')
+        }
         try {
           await models.DepartmentUser.deleteOne({
             userId,
